refactor(vscode-agent-tasks): extract worktree discovery into shared lib
Move findGwRoot, getWorktreePathsFromGw, getWorktreePathsFromGit, and
discoverWorktreePaths from sessions-provider.ts into
src/lib/worktree-discovery.ts so both the Sessions panel and the
upcoming Agent Tasks multi-worktree support can share them.
